delete some unnecessary folders and add database create script
