+++ backend/app/index.php
@@ -1,4 +1,6 @@
 <?php
     // Include Composer autoloader to load dependencies
     require_once __DIR__ . '/../vendor/autoload.php';
+
+
 ?>
