--- backend/app/models/user.php
@@ -1,7 +1,0 @@
-<?php
-    namespace User\Models;
-    use Illuminate\Database\Eloquent\Model;
-    class Product extends Model {
-        protected $fillable = ['name', 'price', 'description'];
-    }
-?>
