--- backend/app/routes/userRoutes.php
@@ -1,23 +1,0 @@
-<?php
-    namespace User\Routes;
-
-    use Symfony\Component\HttpFoundation\Request; // Import the Request class
-    use Symfony\Component\Routing\Route; // Import the Route class
-
-    class UserRouter{
-        private $router;
-
-        public function __construct() {
-            // $this->router = new Route();
-        }
-
-        public function defineRoutes() {
-            // Define your routes here, as shown in the previous response
-            // Route::get('/users', 'UserController@index');
-            // Route::post('/users', 'UserController@store');
-            // Route::get('/users/{id}', 'UserController@show');
-            // Route::put('/users/{id}', 'UserController@update');
-            // Route::delete('/users/{id}', 'UserController@destroy');
-        }
-    }
-?>
