--- backend/public/index.php
@@ -1,23 +1,0 @@
-<?php
-
-    // Autoload Composer dependencies
-    require __DIR__ . '/../vendor/autoload.php';
-
-    // Include any additional initialization files if needed
-    require __DIR__ . '/../config/db.php'; // Example: Database configuration
-
-    // Create a Symfony HTTP request
-    $request = \Symfony\Component\HttpFoundation\Request::createFromGlobals();
-
-    // Your application logic goes here
-    // For example, you might use a router to handle the request
-
-    // Define your routes and dispatch the request
-    $router = new YourNamespace\Router(); // Adjust the namespace based on your project structure
-    $router->defineRoutes();
-    $response = $router->dispatch($request);
-
-    // Send the response to the client
-    $response->send();
-
-?>
